backend/requirements.txt
fastapi
uvicorn
websockets
pydantic
python-dotenv
sqlmodel
asyncpg
alembic
bleach